+++ src/Server/Persistence/Systems/Persistence/PersistenceSystem.cs
@@ -84,7 +84,9 @@ public sealed class PersistenceSystem : ISystem
         EventId.Server_Persistence_RetrieveEntity,
         EventId.Server_Persistence_RetrieveWorldSegment,
         EventId.Server_Persistence_Synchronize,
-        EventId.Server_Accounts_SelectPlayer
+        EventId.Server_Accounts_SelectPlayer,
+        EventId.Core_Data_GlobalSet,
+        EventId.Core_Data_GlobalRemoved
     };
 
     public int WorkloadEstimate => 20;